Change isTouchDevice to return false for TouchScreen laptops
I discovered that when you play TowerStorm on a touch screen laptop
the method window.ontouchstart exists and so it acts as if you're
playing on a touch-screen when really most players would be using a mouse.

Changing it to specifically detect common tablets and phones in the User Agent
instead, this may make it not detect some rare tables but it can be extended
later to support them.